requirements.txt
boto3==1.28.23
snowflake-connector-python==3.1.1
python-dotenv
pympler